+++ CHANGELOG.md
@@ -45,6 +45,7 @@
     `SerializationException` or `NotImplemented` error
     * `CallbackGame` currently is an object
         * It is possible to use `CallbackGame` for now
+            * `CallbackGameInlineKeyboardButton` now will not accept `callbackGame` as income object
     * Now it is possible to pass exception handler in webhook
 
 ### 0.26.3